pyproject.toml
[project]
name = "zeromq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "msgspec>=0.19.0",
    "pydantic>=2.10.4",
    "pyzmq>=26.2.0",
]
